requirements.txt
streamlit
google-genai
streamlit-lottie
requests
gTTS
pandas
fpdf2
st-gsheets-connection
streamlit-mic-recorder
SpeechRecognition
librosa
soundfile